Applied sane coding standard Moved most types into separate files Renamed interfaces where it made sense
