+++ Randomizer.SMZ3/EmbeddedStream.cs
@@ -1,16 +1,15 @@
 ﻿using System.IO;
 using System.Reflection;
 
-namespace Randomizer.SMZ3 {
-
-    public class EmbeddedStream {
-
-        public static Stream For(string name) {
+namespace Randomizer.SMZ3
+{
+    public class EmbeddedStream
+    {
+        public static Stream For(string name)
+        {
             var type = typeof(EmbeddedStream);
             var assembly = Assembly.GetAssembly(type);
             return assembly.GetManifestResourceStream($"{type.Namespace}.{name}");
         }
-
     }
-
 }
