+++ Randomizer.SMZ3/Location.cs
@@ -2,26 +2,14 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Randomizer.SMZ3 {
+namespace Randomizer.SMZ3
+{
+    public delegate bool Requirement(Progression items);
 
-    enum LocationType {
-        Regular,
-        HeraStandingKey,
-        Pedestal,
-        Ether,
-        Bombos,
-        NotInDungeon,
+    public delegate bool Verification(Item item, Progression items);
 
-        Visible,
-        Chozo,
-        Hidden
-    }
-
-    delegate bool Requirement(Progression items);
-    delegate bool Verification(Item item, Progression items);
-
-    class Location {
-
+    public class Location
+    {
         public int Id { get; set; }
         public string Name { get; set; }
         public LocationType Type { get; set; }
@@ -29,23 +17,26 @@ namespace Randomizer.SMZ3 {
         public Item Item { get; set; }
         public Region Region { get; set; }
 
-        public int Weight {
+        public int Weight
+        {
             get { return weight ?? Region.Weight; }
         }
 
-        readonly Requirement canAccess;
-        Verification alwaysAllow;
-        Verification allow;
-        int? weight;
+        private readonly Requirement canAccess;
+        private Verification alwaysAllow;
+        private Verification allow;
+        private int? weight;
 
         public bool ItemIs(ItemType type, World world) => Item?.Is(type, world) ?? false;
         public bool ItemIsNot(ItemType type, World world) => !ItemIs(type, world);
 
         public Location(Region region, int id, int address, LocationType type, string name)
-            : this(region, id, address, type, name, items => true) {
+            : this(region, id, address, type, name, items => true)
+        {
         }
 
-        public Location(Region region, int id, int address, LocationType type, string name, Requirement access) {
+        public Location(Region region, int id, int address, LocationType type, string name, Requirement access)
+        {
             Region = region;
             Id = id;
             Name = name;
@@ -56,26 +47,31 @@ namespace Randomizer.SMZ3 {
             allow = (item, items) => true;
         }
 
-        public Location Weighted(int? weight) {
+        public Location Weighted(int? weight)
+        {
             this.weight = weight;
             return this;
         }
 
-        public Location AlwaysAllow(Verification allow) {
+        public Location AlwaysAllow(Verification allow)
+        {
             alwaysAllow = allow;
             return this;
         }
 
-        public Location Allow(Verification allow) {
+        public Location Allow(Verification allow)
+        {
             this.allow = allow;
             return this;
         }
 
-        public bool Available(Progression items) {
+        public bool Available(Progression items)
+        {
             return Region.CanEnter(items) && canAccess(items);
         }
 
-        public bool CanFill(Item item, Progression items) {
+        public bool CanFill(Item item, Progression items)
+        {
             var oldItem = Item;
             Item = item;
             bool fillable = alwaysAllow(item, items) || (Region.CanFill(item, items) && allow(item, items) && Available(items));
@@ -84,34 +80,41 @@ namespace Randomizer.SMZ3 {
         }
     }
 
-    static class LocationsExtensions {
+    internal static class LocationsExtensions
+    {
 
-        public static Location Get(this IEnumerable<Location> locations, string name) {
+        public static Location Get(this IEnumerable<Location> locations, string name)
+        {
             var location = locations.FirstOrDefault(l => l.Name == name);
             if (location == null)
                 throw new ArgumentException($"Could not find location name {name}", nameof(name));
             return location;
         }
 
-        public static IEnumerable<Location> Empty(this IEnumerable<Location> locations) {
+        public static IEnumerable<Location> Empty(this IEnumerable<Location> locations)
+        {
             return locations.Where(l => l.Item == null);
         }
 
-        public static IEnumerable<Location> Filled(this IEnumerable<Location> locations) {
+        public static IEnumerable<Location> Filled(this IEnumerable<Location> locations)
+        {
             return locations.Where(l => l.Item != null);
         }
 
-        public static IEnumerable<Location> AvailableWithinWorld(this IEnumerable<Location> locations, IEnumerable<Item> items) {
+        public static IEnumerable<Location> AvailableWithinWorld(this IEnumerable<Location> locations, IEnumerable<Item> items)
+        {
             return locations.Select(x => x.Region.World).Distinct().SelectMany(world =>
                 locations.Where(l => l.Region.World == world).Available(items.Where(i => i.World == world)));
         }
 
-        public static IEnumerable<Location> Available(this IEnumerable<Location> locations, IEnumerable<Item> items) {
+        public static IEnumerable<Location> Available(this IEnumerable<Location> locations, IEnumerable<Item> items)
+        {
             var progression = new Progression(items);
             return locations.Where(l => l.Available(progression));
         }
 
-        public static IEnumerable<Location> CanFillWithinWorld(this IEnumerable<Location> locations, Item item, IEnumerable<Item> items) {
+        public static IEnumerable<Location> CanFillWithinWorld(this IEnumerable<Location> locations, Item item, IEnumerable<Item> items)
+        {
             var itemWorldProgression = new Progression(items.Where(i => i.World == item.World).Append(item));
             var worldProgression = locations.Select(x => x.Region.World).Distinct().ToDictionary(world => world.Id, world => new Progression(items.Where(i => i.World == world)));
