+++ Randomizer.SMZ3/LocationType.cs
@@ -1,0 +1,17 @@
+﻿namespace Randomizer.SMZ3
+{
+    public enum LocationType
+    {
+        Regular,
+        HeraStandingKey,
+        Pedestal,
+        Ether,
+        Bombos,
+        NotInDungeon,
+
+        Visible,
+        Chozo,
+        Hidden
+    }
+
+}
