+++ Randomizer.SMZ3/Randomizer.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Threading;
+
 using Randomizer.Shared.Contracts;
 
-namespace Randomizer.SMZ3 {
-
-    public class Randomizer : IRandomizer {
-
+namespace Randomizer.SMZ3
+{
+    public class Randomizer : IRandomizer
+    {
         public static readonly Version version = new Version(11, 2);
 
         public string Id => "smz3";
@@ -16,16 +17,20 @@ namespace Randomizer.SMZ3 {
         public string Version => version.ToString();
         public List<IRandomizerOption> Options => RandomizerOptions.List;
 
-        static readonly Regex legalCharacters = new Regex(@"[A-Z0-9]", RegexOptions.IgnoreCase);
-        static readonly Regex illegalCharacters = new Regex(@"[^A-Z0-9]", RegexOptions.IgnoreCase);
-        static readonly Regex continousSpace = new Regex(@" +");
+        private static readonly Regex legalCharacters = new Regex(@"[A-Z0-9]", RegexOptions.IgnoreCase);
+        private static readonly Regex illegalCharacters = new Regex(@"[^A-Z0-9]", RegexOptions.IgnoreCase);
+        private static readonly Regex continousSpace = new Regex(@" +");
 
-        public ISeedData GenerateSeed(IDictionary<string, string> options, string seed, CancellationToken cancellationToken) {
+        public ISeedData GenerateSeed(IDictionary<string, string> options, string seed, CancellationToken cancellationToken)
+        {
             int randoSeed;
-            if (string.IsNullOrEmpty(seed)) {
+            if (string.IsNullOrEmpty(seed))
+            {
                 randoSeed = System.Security.Cryptography.RandomNumberGenerator.GetInt32(0, int.MaxValue);
                 seed = randoSeed.ToString();
-            } else {
+            }
+            else
+            {
                 randoSeed = int.Parse(seed);
                 /* The Random ctor takes the absolute value of a negative seed.
                  * This is an non-obvious behavior so we treat a negative value
@@ -38,17 +43,21 @@ namespace Randomizer.SMZ3 {
             var config = RandomizerOptions.Parse(options);
 
             /* FIXME: Just here to semi-obfuscate race seeds until a better solution is in place */
-            if (config.Race) {
+            if (config.Race)
+            {
                 randoRnd = new Random(randoRnd.Next());
             }
 
             var worlds = new List<World>();
-            if (config.SingleWorld) {
+            if (config.SingleWorld)
+            {
                 worlds.Add(new World(config, "Player", 0, new HexGuid()));
             }
-            else {
+            else
+            {
                 var players = options.ContainsKey("players") ? int.Parse(options["players"]) : 1;
-                for (var p = 0; p < players; p++) {
+                for (var p = 0; p < players; p++)
+                {
                     var found = options.TryGetValue($"player-{p}", out var player);
                     if (!found)
                         throw new ArgumentException($"No name provided for player {p + 1}");
@@ -65,7 +74,8 @@ namespace Randomizer.SMZ3 {
             var playthrough = new Playthrough(worlds, config);
             var spheres = playthrough.Generate();
 
-            var seedData = new SeedData {
+            var seedData = new SeedData
+            {
                 Guid = new HexGuid(),
                 Seed = seed,
                 Game = Name,
@@ -76,11 +86,13 @@ namespace Randomizer.SMZ3 {
             };
 
             /* Make sure RNG is the same when applying patches to the ROM to have consistent RNG for seed identifer etc */
-            int patchSeed = randoRnd.Next();
-            foreach (var world in worlds) {
+            var patchSeed = randoRnd.Next();
+            foreach (var world in worlds)
+            {
                 var patchRnd = new Random(patchSeed);
                 var patch = new Patch(world, worlds, seedData.Guid, config.Race ? 0 : randoSeed, patchRnd);
-                var worldData = new WorldData {
+                var worldData = new WorldData
+                {
                     Id = world.Id,
                     Guid = world.Guid,
                     Player = world.Player,
@@ -96,9 +108,10 @@ namespace Randomizer.SMZ3 {
             return seedData;
         }
 
-        public Dictionary<int, ILocationTypeData> GetLocations() => 
+        public Dictionary<int, ILocationTypeData> GetLocations() =>
             new World(new Config(), "", 0, "")
-                .Locations.Select(location => new LocationTypeData {
+                .Locations.Select(location => new LocationTypeData
+                {
                     Id = location.Id,
                     Name = location.Name,
                     Type = location.Type.ToString(),
@@ -107,12 +120,14 @@ namespace Randomizer.SMZ3 {
                 }).Cast<ILocationTypeData>().ToDictionary(locationData => locationData.Id);
 
         public Dictionary<int, IItemTypeData> GetItems() =>
-            Enum.GetValues(typeof(ItemType)).Cast<ItemType>().Select(i => new ItemTypeData {
+            Enum.GetValues(typeof(ItemType)).Cast<ItemType>().Select(i => new ItemTypeData
+            {
                 Id = (int)i,
                 Name = i.GetDescription()
             }).Cast<IItemTypeData>().ToDictionary(itemTypeData => itemTypeData.Id);
 
-        static string CleanPlayerName(string name) {
+        private static string CleanPlayerName(string name)
+        {
             name = illegalCharacters.Replace(name, " ");
             name = continousSpace.Replace(name, " ");
             return name.Trim();
@@ -120,7 +135,8 @@ namespace Randomizer.SMZ3 {
 
     }
 
-    public class RandomizerOption : IRandomizerOption {
+    public class RandomizerOption : IRandomizerOption
+    {
         public string Key { get; set; }
         public string Description { get; set; }
         public RandomizerOptionType Type { get; set; }
@@ -128,7 +144,8 @@ namespace Randomizer.SMZ3 {
         public string Default { get; set; }
     }
 
-    public class SeedData : ISeedData {
+    public class SeedData : ISeedData
+    {
         public string Guid { get; set; }
         public string Seed { get; set; }
         public string Game { get; set; }
@@ -138,7 +155,8 @@ namespace Randomizer.SMZ3 {
         public List<Dictionary<string, string>> Playthrough { get; set; }
     }
 
-    public class WorldData : IWorldData {
+    public class WorldData : IWorldData
+    {
         public int Id { get; set; }
         public string Guid { get; set; }
         public string Player { get; set; }
@@ -146,18 +164,21 @@ namespace Randomizer.SMZ3 {
         public List<ILocationData> Locations { get; set; }
     }
 
-    public class LocationData : ILocationData {
+    public class LocationData : ILocationData
+    {
         public int LocationId { get; set; }
         public int ItemId { get; set; }
         public int ItemWorldId { get; set; }
     }
 
-    public class ItemTypeData : IItemTypeData {
+    public class ItemTypeData : IItemTypeData
+    {
         public int Id { get; set; }
         public string Name { get; set; }
     }
 
-    public class LocationTypeData : ILocationTypeData {
+    public class LocationTypeData : ILocationTypeData
+    {
         public int Id { get; set; }
         public string Name { get; set; }
         public string Type { get; set; }
