+++ Randomizer.SMZ3/Region.cs
@@ -2,46 +2,10 @@
 using System.ComponentModel;
 using System.Linq;
 
-namespace Randomizer.SMZ3 {
-
-    enum RewardType {
-        [Description("None")]
-        None,
-        [Description("Agahnim")]
-        Agahnim,
-        [Description("Green Pendant")]
-        PendantGreen,
-        [Description("Blue/Red Pendant")]
-        PendantNonGreen,
-        [Description("Blue Crystal")]
-        CrystalBlue,
-        [Description("Red Crystal")]
-        CrystalRed,
-        [Description("Golden Four Boss")]
-        GoldenFourBoss
-    }
-
-    interface IReward {
-        RewardType Reward { get; set; }
-        bool CanComplete(Progression items);
-    }
-
-    interface IMedallionAccess {
-        ItemType Medallion { get; set; }
-    }
-
-    abstract class SMRegion : Region {
-        public SMLogic Logic => Config.SMLogic;
-        public SMRegion(World world, Config config) : base(world, config) { }
-    }
-
-    abstract class Z3Region : Region {
-        public Z3Region(World world, Config config)
-            : base(world, config) { }
-    }
-
-    abstract class Region {
-
+namespace Randomizer.SMZ3
+{
+    public abstract class Region
+    {
         public virtual string Name { get; }
         public virtual string Area => Name;
 
@@ -55,25 +19,30 @@ namespace Randomizer.SMZ3 {
         private Dictionary<string, Location> locationLookup { get; set; }
         public Location GetLocation(string name) => locationLookup[name];
 
-        public Region(World world, Config config) {
+        protected Region(World world, Config config)
+        {
             Config = config;
             World = world;
             locationLookup = new Dictionary<string, Location>();
         }
 
-        public void GenerateLocationLookup() {
+        public void GenerateLocationLookup()
+        {
             locationLookup = Locations.ToDictionary(l => l.Name, l => l);
         }
 
-        public bool IsRegionItem(Item item) {
+        public bool IsRegionItem(Item item)
+        {
             return RegionItems.Contains(item.Type);
         }
 
-        public virtual bool CanFill(Item item, Progression items) {
+        public virtual bool CanFill(Item item, Progression items)
+        {
             return Config.Keysanity || !item.IsDungeonItem || IsRegionItem(item);
         }
 
-        public virtual bool CanEnter(Progression items) {
+        public virtual bool CanEnter(Progression items)
+        {
             return true;
         }
