+++ Randomizer.SMZ3/Text/Texts.cs
@@ -2,21 +2,25 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+
 using Randomizer.SMZ3.Regions.Zelda;
+
 using SharpYaml.Model;
+
 using static Randomizer.SMZ3.ItemType;
 
-namespace Randomizer.SMZ3.Text {
+namespace Randomizer.SMZ3.Text
+{
+    public static class Texts
+    {
+        private static readonly YamlMapping scripts;
+        private static readonly IList<string> blind;
+        private static readonly IList<string> ganon;
+        private static readonly IList<string> tavernMan;
+        private static readonly IList<string> triforceRoom;
 
-    class Texts {
-
-        static YamlMapping scripts;
-        static IList<string> blind;
-        static IList<string> ganon;
-        static IList<string> tavernMan;
-        static IList<string> triforceRoom;
-
-        static Texts() {
+        static Texts()
+        {
             scripts = ParseYamlScripts("Text.Scripts.General.yaml") as YamlMapping;
             blind = ParseTextScript("Text.Scripts.BlindExtra.txt");
             ganon = ParseTextScript("Text.Scripts.GanonExtra.txt");
@@ -24,13 +28,15 @@ namespace Randomizer.SMZ3.Text {
             triforceRoom = ParseTextScript("Text.Scripts.TriforceRoomExtra.txt");
         }
 
-        static YamlElement ParseYamlScripts(string resource) {
+        private static YamlElement ParseYamlScripts(string resource)
+        {
             using var stream = EmbeddedStream.For(resource);
             using var reader = new StreamReader(stream);
             return YamlStream.Load(reader).First().Contents;
         }
 
-        static IList<string> ParseTextScript(string resource) {
+        private static IList<string> ParseTextScript(string resource)
+        {
             using var stream = EmbeddedStream.For(resource);
             using var reader = new StreamReader(stream);
             return reader.ReadToEnd().Replace("\r", "").Split("---\n")
@@ -39,35 +45,41 @@ namespace Randomizer.SMZ3.Text {
                 .ToList();
         }
 
-        public static string SahasrahlaReveal(Region dungeon) {
-            string text = (scripts["SahasrahlaReveal"] as YamlValue).Value;
+        public static string SahasrahlaReveal(Region dungeon)
+        {
+            var text = (scripts["SahasrahlaReveal"] as YamlValue).Value;
             return text.Replace("<dungeon>", dungeon.Area);
         }
 
-        public static string BombShopReveal(IEnumerable<Region> dungeons) {
+        public static string BombShopReveal(IEnumerable<Region> dungeons)
+        {
             var (first, second, _) = dungeons;
-            string text = (scripts["BombShopReveal"] as YamlValue).Value;
+            var text = (scripts["BombShopReveal"] as YamlValue).Value;
             return text.Replace("<first>", first.Area).Replace("<second>", second.Area);
         }
 
-        public static string GanonThirdPhaseSingle(Region silvers) {
+        public static string GanonThirdPhaseSingle(Region silvers)
+        {
             var node = (scripts["GanonSilversReveal"] as YamlMapping)["single"] as YamlMapping;
-            string text = (node[silvers is GanonsTower ? "local" : "remote"] as YamlValue).Value;
+            var text = (node[silvers is GanonsTower ? "local" : "remote"] as YamlValue).Value;
             return text.Replace("<region>", silvers.Area);
         }
 
-        public static string GanonThirdPhaseMulti(Region silvers, World myWorld) {
+        public static string GanonThirdPhaseMulti(Region silvers, World myWorld)
+        {
             var node = (scripts["GanonSilversReveal"] as YamlMapping)["multi"] as YamlMapping;
             if (silvers.World == myWorld)
                 return (node["local"] as YamlValue).Value;
             var player = silvers.World.Player;
             player = player.PadLeft(7 + player.Length / 2);
-            string text = (node["remote"] as YamlValue).Value;
+            var text = (node["remote"] as YamlValue).Value;
             return text.Replace("<player>", player);
         }
 
-        public static string ItemTextbox(Item item) {
-            var name = item.Type switch {
+        public static string ItemTextbox(Item item)
+        {
+            var name = item.Type switch
+            {
                 _ when item.IsMap => "Map",
                 _ when item.IsCompass => "Compass",
                 _ when item.IsKeycard => "Keycard",
@@ -97,7 +109,7 @@ namespace Randomizer.SMZ3.Text {
 
         public static string TriforceRoom(Random rnd) => RandomLine(rnd, triforceRoom);
 
-        static string RandomLine(Random rnd, IList<string> lines) => lines[rnd.Next(lines.Count)];
+        private static string RandomLine(Random rnd, IList<string> lines) => lines[rnd.Next(lines.Count)];
 
     }
