+++ Randomizer.SMZ3/Z3Region.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using System.ComponentModel;
+using System.Linq;
+
+namespace Randomizer.SMZ3
+{
+
+    public abstract class Z3Region : Region {
+        public Z3Region(World world, Config config)
+            : base(world, config) { }
+    }
+
+}
